requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
ccxt==4.1.77
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
ta==0.10.2
schedule==1.2.0 